refactor: Extraction, Alignment, and History Logic
Moving logic outside of price chart so that it only represents the updating of the chart
