+++ frontend/src/types/price-point.ts
@@ -1,0 +1,5 @@
+export interface PricePoint {
+    timestamp: number;
+    price: number;
+    source: string
+  }
